enhancement/1349 - moves cid to aircraftmodel (breaks tests)
