+++ src/assets/scripts/client/aircraft/AircraftModel.js
@@ -76,6 +76,7 @@ import {
     INVALID_NUMBER,
     TIME
 } from '../constants/globalConstants';
+import { leftPad } from '../utilities/generalUtilities';
 
 /**
  * @property FLIGHT_RULES
@@ -437,6 +438,31 @@ export default class AircraftModel {
          */
         this.isRemovable = false;
 
+        /**
+         * This value is generated by the `AircraftController` and is
+         * unique to every aircraft
+         *
+         * Currently only numeric, but could be made alpha-numberic
+         *
+         * Padded with leading zeros in `#cid` for display purposes
+         *
+         * @property cidValue
+         * @type {number}
+         * @default 0
+         */
+        this.cidValue = 0;
+
+        /**
+         * A string version of `#cidValue` for display purposes
+         * Has been padded with leading zeros so that it is always
+         * displayed as three digits
+         *
+         * @property cid
+         * @type {string}
+         * @default ''
+         */
+        this.cid = '';
+
         // Set to true when simulating future movements of the aircraft
         // Should be checked before updating global state such as score
         // or HTML.
@@ -608,6 +634,8 @@ export default class AircraftModel {
         this.airlineId = data.airline;
         this.airlineCallsign = data.airlineCallsign;
         this.flightNumber = data.callsign;
+        this.cidValue = data.cid;
+        this.cid = leftPad(this.cidValue, 3);
         this.category = data.category;
         this.heading = data.heading;
         this.altitude = data.altitude;
@@ -652,6 +680,7 @@ export default class AircraftModel {
             id: this.id,
             insideCenter: this.isControllable,
             callsign: this.callsign,
+            cid: this.cid,
             transponderCode: this.transponderCode,
             icaoWithWeightClass: this.model.icaoWithWeightClass,
             assignedAltitude,
