+++ src/assets/scripts/client/aircraft/StripView/StripViewController.js
@@ -3,17 +3,7 @@ import _random from 'lodash/random';
 import _without from 'lodash/without';
 import StripViewCollection from './StripViewCollection';
 import StripViewModel from './StripViewModel';
-import { INVALID_INDEX } from '../../constants/globalConstants';
 import { SELECTORS } from '../../constants/selectors';
-
-/**
- * The highest number allowed for a cid value
- *
- * @property CID_UPPER_BOUND
- * @type {number}
- * @final
- */
-const CID_UPPER_BOUND = 999;
 
 /**
  * Control modifications of the `$stripViewList` and coordinate
@@ -68,13 +58,6 @@ export default class StripViewController {
          * @type {JQuery|HTMLElement}
          */
         this.$stripListTrigger = $(SELECTORS.DOM_SELECTORS.STRIP_VIEW_TRIGGER);
-
-        /**
-         * @property _cidNumbersInUse
-         * @type {array<number>}
-         * @private
-         */
-        this._cidNumbersInUse = [];
 
         return this._init()
             .enable();
@@ -174,8 +157,7 @@ export default class StripViewController {
      * @return {StripViewModel}
      */
     createStripView(aircraftModel) {
-        const stripViewCid = this._generateCidNumber();
-        const stripViewModel = new StripViewModel(aircraftModel, stripViewCid);
+        const stripViewModel = new StripViewModel(aircraftModel);
 
         this._collection.addItem(stripViewModel);
 
@@ -261,7 +243,6 @@ export default class StripViewController {
             return;
         }
 
-        this._removeCidFromUse(stripViewModel.cid);
         this._collection.removeItem(stripViewModel);
         stripViewModel.destroy();
     }
@@ -314,46 +295,4 @@ export default class StripViewController {
      */
     // eslint-disable-next-line no-unused-vars
     _onStripListClickOutsideStripViewModel = (event) => this.findAndDeselectActiveStripView();
-
-    /**
-     * Generate a unique number to represent a `CID`
-     *
-     * Should be displayed with leading zeros, so a `CID` value of `1` should be displayed as `001`
-     *
-     * @for StripViewController
-     * @method _generateCidNumber
-     * @return nextCid {number}
-     * @private
-     */
-    _generateCidNumber() {
-        const nextCid = _random(1, CID_UPPER_BOUND);
-
-        if (this._cidNumbersInUse.indexOf(nextCid) !== INVALID_INDEX) {
-            return this._generateCidNumber();
-        }
-
-        this._cidNumbersInUse.push(nextCid);
-
-        return nextCid;
-    }
-
-    /**
-     * Remove a given `#cid` from use
-     *
-     * Used when an aircraft has landed or departed controlled airspace
-     *
-     * @for StripViewController
-     * @method _removeCidFromUse
-     * @param cid {number}
-     * @private
-     */
-    _removeCidFromUse(cid) {
-        const cidIndex = this._cidNumbersInUse.indexOf(cid);
-
-        if (cidIndex === INVALID_INDEX) {
-            return;
-        }
-
-        this._cidNumbersInUse = _without(this._cidNumbersInUse, cid);
-    }
 }
