+++ src/assets/scripts/client/aircraft/StripView/StripViewModel.js
@@ -28,9 +28,8 @@ export default class StripViewModel extends BaseModel {
     /**
      * @constructor
      * @param aircraftModel {object}
-     * @param cidValue {number}
      */
-    constructor(aircraftModel, cidValue) {
+    constructor(aircraftModel) {
         super('stripViewModel');
 
         /**
@@ -150,29 +149,14 @@ export default class StripViewModel extends BaseModel {
         this.$aircraftTypeView = null;
 
         /**
-         * This value is generated by the `StripViewController` and is
-         * unique to every `StripViewModel`
-         *
-         * Currently only numeric, but could be made alpha-numberic
-         *
-         * Padded with leading zeros in `#cidString` for display purposes
+         * Aircraft CID
          *
          * @property cid
-         * @type {number}
-         * @default cidValue
+         * @type {String}
+         * @default ''
+         * @private
          */
-        this.cid = cidValue;
-
-        /**
-         * A string version of `#cid` for display purposes
-         * Has been padded with leading zeros so that it is always
-         * displayed as three digits
-         *
-         * @property cidString
-         * @type {string}
-         * @default cidValue
-         */
-        this.cidString = leftPad(cidValue, 3);
+        this._cid = '';
 
         /**
          * HTML Element that holds `#_cidValue`
@@ -389,6 +373,7 @@ export default class StripViewModel extends BaseModel {
         const {
             insideCenter,
             callsign,
+            cid,
             icaoWithWeightClass,
             transponderCode,
             assignedAltitude,
@@ -400,6 +385,7 @@ export default class StripViewModel extends BaseModel {
 
         this.insideCenter = insideCenter;
         this._callsign = callsign;
+        this._cid = cid;
         this._transponder = transponderCode;
         this._aircraftType = icaoWithWeightClass;
         this._assignedAltitude = assignedAltitude;
@@ -506,7 +492,7 @@ export default class StripViewModel extends BaseModel {
     _redraw() {
         this.$callsignView.text(this._callsign);
         this.$aircraftTypeView.text(this._aircraftType);
-        this.$cidView.text(this.cidString);
+        this.$cidView.text(this._cid);
         this.$transponderView.text(this._transponder);
         this.$assignedAltitudeView.text(this._assignedAltitude);
         this.$flightPlanAltitudeView.text(this._flightPlanAltitude);
@@ -554,8 +540,7 @@ export default class StripViewModel extends BaseModel {
         this.$callsignView = null;
         this._aircraftType = '';
         this.$aircraftTypeView = null;
-        this.cid = INVALID_NUMBER;
-        this.cidString = '';
+        this._cid = '';
         this.$cidView = null;
         this._transponder = 1200;
         this.$transponderView = null;
